Вынес Extension из класса WatchCinemaDbContext в отдельный класс.
